requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
docker==2.5.1
docker-pycreds==0.2.3
idna==2.6
kombu==4.1.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
websocket-client==0.47.0
